Add selectTabCombinedFilters to unify selecting tab filters and remove dependencies on global services, which can trigger when switching tabs
