+++ src/platform/plugins/shared/discover/public/application/main/components/top_nav/discover_topnav.tsx
@@ -16,16 +16,9 @@ import {
 } from '@kbn/discover-utils';
 import type { ESQLEditorRestorableState } from '@kbn/esql-editor';
 import { useESQLQueryStats } from '@kbn/esql/public';
-import {
-  type Filter,
-  type Query,
-  type TimeRange,
-  type AggregateQuery,
-  isOfAggregateQueryType,
-} from '@kbn/es-query';
+import { type Query, type TimeRange, type AggregateQuery } from '@kbn/es-query';
 import type { DataViewPickerProps, UnifiedSearchDraft } from '@kbn/unified-search-plugin/public';
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import { cloneDeep } from 'lodash';
 import { ESQL_TRANSITION_MODAL_KEY } from '../../../../../common/constants';
 import { useDiscoverCustomization } from '../../../../customizations';
 import { useDiscoverServices } from '../../../../hooks/use_discover_services';
@@ -33,6 +26,7 @@ import { useIsEsqlMode } from '../../hooks/use_is_esql_mode';
 import type { DiscoverStateContainer } from '../../state_management/discover_state';
 import {
   internalStateActions,
+  selectTabCombinedFilters,
   useAppStateSelector,
   useCurrentDataView,
   useCurrentTabAction,
@@ -47,8 +41,6 @@ import { onSaveDiscoverSession } from './save_discover_session';
 import { useDiscoverTopNav } from './use_discover_topnav';
 import { useESQLVariables } from './use_esql_variables';
 import type { UpdateESQLQueryFn } from '../../../../context_awareness/types';
-
-const EMPTY_FILTERS: Filter[] = [];
 
 export interface DiscoverTopNavProps {
   savedQuery?: string;
@@ -82,16 +74,7 @@ export const DiscoverTopNav = ({
   const { timeRangeAbsolute } = useCurrentTabSelector((tab) => tab.dataRequestParams);
   const refreshInterval = useCurrentTabSelector((state) => state.globalState.refreshInterval);
   const timeRangeRelative = useCurrentTabSelector((state) => state.globalState.timeRange);
-  const appFilters = useAppStateSelector((state) => state.filters);
-  const globalFilters = useCurrentTabSelector((state) => state.globalState.filters);
-
-  const filtersMemoized = useMemo(() => {
-    if (isOfAggregateQueryType(query)) {
-      return EMPTY_FILTERS;
-    }
-    const allFilters = [...(globalFilters ?? []), ...(appFilters ?? [])];
-    return allFilters.length ? cloneDeep(allFilters) : EMPTY_FILTERS;
-  }, [appFilters, globalFilters, query]);
+  const filters = useCurrentTabSelector(selectTabCombinedFilters);
 
   const { savedDataViews, adHocDataViews } = useDataViewsForPicker();
   const dataView = useCurrentDataView();
@@ -362,7 +345,7 @@ export const DiscoverTopNav = ({
         onSavedQueryIdChange={updateSavedQueryId}
         disableSubscribingToGlobalDataServices={true}
         query={query}
-        filters={filtersMemoized}
+        filters={filters}
         dateRangeFrom={timeRangeRelative?.from}
         dateRangeTo={timeRangeRelative?.to}
         refreshInterval={refreshInterval?.value}
