+++ src/platform/plugins/shared/discover/public/application/main/state_management/redux/index.ts
@@ -48,6 +48,7 @@ export {
   selectRecentlyClosedTabs,
   selectTab,
   selectTabAppState,
+  selectTabCombinedFilters,
   selectIsTabsBarHidden,
   selectHasUnsavedChanges,
 } from './selectors';
