+++ src/platform/plugins/shared/discover/public/application/main/state_management/redux/selectors/index.ts
@@ -12,6 +12,7 @@ export {
   selectRecentlyClosedTabs,
   selectTab,
   selectTabAppState,
+  selectTabCombinedFilters,
   selectIsTabsBarHidden,
 } from './tabs';
 export { type HasUnsavedChangesResult, selectHasUnsavedChanges } from './unsaved_changes';
